pp-56: friend request functionality

+++ client/src/components/Friend/friend.js
@@ -43,18 +43,40 @@ const Friend = () => {
     }
   
     try {
-      const response = await axios.post(`/addFriend/${userId}`, {
-        friendId: userIdToAdd,
+      // Check if the user you are trying to add is already in your friend list
+      const response = await axios.get(`/checkFriendStatus/${userId}/${userIdToAdd}`);
+  
+      if (response.data === 'friends') {
+        console.log('You are already friends');
+        alert('You are already friends');
+        return;
+      }
+      if (response.data === 'in_request') {
+        console.log('The other person already sent you friend request');
+        alert('The other person already sent you friend request');
+        return;
+      }
+      
+      // Send a friend request to the user with userIdToAdd
+      const response2 = await axios.post(`/addFriendRequest/${userIdToAdd}`, {
+        senderId: userId,
       });
-      console.log('Friend added:', response.data);
-      alert("Friend added successfully :)");
+  
+      console.log('Friend request sent:', response2.data);
+      alert('Friend request sent successfully :)');
     } catch (error) {
       console.log(error);
     }
   };
+
   
   const handleDeleteClick = async (friendIdToDelete) => {
-    
+    if (userId === friendIdToDelete) {
+      console.log("Cannot remove yourself as a friend");
+      alert("Cannot remove yourself as a friend");
+      return;
+    }
+  
     try {
       const response = await axios.delete(`/removeFriend/${userId}`, {
         data: { friendId: friendIdToDelete },
@@ -62,11 +84,8 @@ const Friend = () => {
   
       if (response.data.success) {
         console.log('Friend removed:', friendIdToDelete);
-        alert("Friend removed successfully :)");
-        
-        
+        alert('Friend removed successfully :)');
       } else if (response.data.needToAdd) {
-        
         console.log('You need to be friends first');
         alert('You need to be friends first');
       } else {
@@ -76,6 +95,7 @@ const Friend = () => {
       console.log('An error occurred while removing friend:', error);
     }
   };
+  
   
   
 
@@ -89,26 +109,28 @@ const Friend = () => {
           onChange={(event) => setQuery(event.target.value)}
         />
         <div className="search"></div>
-        <div id="users">
-          <ul id="users_ul">
-            {filteredData.map((item, index) => {
-              const { fname, lname, email, _id } = item;
-              const fullName = `${fname} ${lname}`;
+        {query !== '' && ( // Conditional rendering based on query state
+          <div id="users">
+            <ul id="users_ul">
+              {filteredData.map((item, index) => {
+                const { fname, lname, email, _id } = item;
+                const fullName = `${fname} ${lname}`;
 
-              return (
-                <li key={index}>
-                  <div className="name">{fullName}</div>
-                  <div className="email">{email}</div>
-                  <div className="button-container">
-                    <button onClick={() => handleButtonClick(_id)}>Add</button>
-                    <button onClick={() => handleDeleteClick(_id)}>Remove</button>
-                  </div>
-                </li>
-              );
-            })}
-          </ul>
-          <span>Number of users: {filteredData.length}</span>
-        </div>
+                return (
+                  <li key={index}>
+                    <div className="name">{fullName}</div>
+                    <div className="email">{email}</div>
+                    <div className="button-container">
+                      <button onClick={() => handleButtonClick(_id)}>Add</button>
+                      <button onClick={() => handleDeleteClick(_id)}>Remove</button>
+                    </div>
+                  </li>
+                );
+              })}
+            </ul>
+            <span>Number of users: {filteredData.length}</span>
+          </div>
+        )}
       </div>
     </div>
   );
